+++ c/src/storage/table.h
@@ -244,7 +244,7 @@ class ActiveTable {
     public:
         // TODO - make private in the future somehow.
         std::map<std::string, table_column*, std::less<>> columns;
-        std::unordered_map<size_t, TablePermissions>* permissions = nullptr;
+        std::unordered_map<long, TablePermissions>* permissions = nullptr;
         table_header header;
         table_column* header_columns;
         std::string_view name;
